requirements.txt
torch
transformers
accelerate
datasets
python-dotenv
torchaudio
librosa
scikit-learn
pandas
numpy
gradio
wandb
matplotlib
seaborn
evaluate